Review tests and avg rating fixes

+++ src/main/java/org/somuga/service/GameService.java
@@ -171,7 +171,7 @@ public class GameService implements IGameService {
     }
 
     @Override
-    public void updateAverageRating(Game game, Integer newRating) {
+    public void updateAverageRating(Game game, Double newRating) {
         game.setAverageRating(newRating);
         gameRepo.save(game);
     }
